updates files to match changes in prototype kit proposal
