--- app/views/apps/subapp/assets/javascripts/subapp.js
@@ -1,1 +1,0 @@
-console.log('Hello!');
