+++ app/views/apps/subapp/index.html
@@ -1,5 +1,1 @@
 {% extends "layouts/subappindex.html" %}
-
-{# {% block description %}
-	<p>Descrption of the version/iteration goes here.</p>
-{% endblock %} #}
