+++ app/views/apps/subapp/subapp-routes.js
@@ -1,5 +1,3 @@
-const _ = require('lodash')
-
 /**
  * routes function for this particular 'subapp'
  * @method exports
@@ -13,7 +11,6 @@ module.exports = function(router, config) {
   router.all([config.route.page, config.route.root + '**/*'], function(req,res,next){
     
 		/* do stuff to all subapp app pages here */
-		console.log("all pages!");
 		
     next()
   
@@ -25,7 +22,6 @@ module.exports = function(router, config) {
 	// called 'sprint1' then the output would be
 	// /apps/sprint1/views/index
 	router.all(config.route.root + 'index', function(req,res,next){
-		console.log('This is the index page!')
 		next()
 	})
