+++ app/views/apps/subapp/views/index.html
@@ -1,5 +1,6 @@
 {% extends currentApp.filePaths.layoutsDir + "default.html" %}
 
 {% block main_content %}
-	<h1>Paul</h1>
+	<h1 class="heading-medium">Page</h1>
+	<p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Quo consequuntur laboriosam odit accusamus nisi, aperiam facere quibusdam excepturi earum dignissimos! Quidem unde ab sed recusandae ullam facilis tempore modi nostrum.</p>	
 {% endblock %}
