--- app/views/apps/subapp/views/paul/paul.html
@@ -1,6 +1,0 @@
-{% extends "layouts/layout.html" %}
-
-{% block content %}
-	<p>Testing.......</p>
-	{{ serviceName | log }}
-{% endblock %}
