+++ app/views/layouts/layout.html
@@ -1,3 +1,5 @@
+{% import "includes/macros/global.html" as macros %}
+
 {% extends "govuk_template.html" %}
 
 {% block head %}
