+++ app/views/layouts/subappindex.html
@@ -2,22 +2,63 @@
 
 {% block content %}
 
+	{{ currentApp | log }}
+
 	<main id="content">
-		
-		<h1 class="heading-xlarge">{{ currentApp.title }}</h1>
-		
-		<p class="lede">Details of this prototype.</p>
-		
-		{% block description %}
-			<p>The contents of <span class="mono-inline">{% raw %}{% block descrition %}{% endblock %}{% endraw %}</span> will be outputted here</p>
-		{% endblock %}
+			
+	  <div class="grid-row">
+	    <div class="column-two-thirds">
+				
+			<p><a href="/" class="link-back">Back</a></p>
+				
+			<h1 class="heading-xlarge">
+				{{ currentApp.title }}
+				<span class="heading-secondary">Details of this prototype</span>
+			</h1>
+
+			{% block description %}
+				<h2 class="heading-medium">Description</h2>
+				<p>To add your own description edit: <span class="mono-inline">{{ currentApp.filePaths.subAppDir }}/index.html</span>. The contents of the 'description' block will be displayed here.<p>
+				<h3 class="heading-small">For example</h3>
+				<p><div class="mono-block">
+					<div>
+					{% filter escape %}
+					{% raw %}
+					{% block description %}
+					{% endraw %}
+					{% endfilter %}
+					</div>
+					<div>
+					{% filter escape %}
+					{% raw %}
+						<p>Your content</p>
+					{% endraw %}
+					{% endfilter %}
+					</div>
+					<div>
+					{% filter escape %}
+					{% raw %}
+						{% endblock %}
+					{% endraw %}
+					{% endfilter %}
+					</div>
+				</div></p>
+			{% endblock %}
+			
+			<br>
+			
+			<p><a href="{{ currentApp.startURL | d('views/') }}" class="button button-start">Launch {{ (currentApp.title) if currentApp.title }} prototype</a></p>
+
+	    </div>
+	  </div>
 		
 		<h3 class="heading-medium">Technical details</h3>
 		
 		<ul class="list-bullet list-block">
 			<li>Node/Express routes file location: <span class="mono-inline">{{ currentApp.filePaths.routesFile }}</span></li>
-			<li>Layouts are located: <span class="mono-inline">{{ currentApp.filePaths.layoutsDir }}</span></li>
+			<li>Sub-application configuration file is located: <span class="mono-inline">{{ currentApp.filePaths.configFile }}</span></li>
 			<li>'Sub application' specific data can be accessed in the templates with <span class="mono-inline">{% raw %}{{ currentApp }}{% endraw %}</span></li>
+			<li>Layouts are located: <span class="mono-inline">{{ currentApp.filePaths.layoutsDir }}</span></li>
 			<li>Session data can be accessed in the templates with <span class="mono-inline">{% raw %}{{ session }}{% endraw %}</span></li>
 			<li>'Post' data can be accessed in the templates with <span class="mono-inline">{% raw %}{{ postData }}{% endraw %}</span></li>
 		</ul>
