--- project.cson
@@ -1,11 +1,0 @@
-{
-  title: "FES Prototype"
-  group: "DWP"
-  paths: [
-    "~/Sites/govuk/DWP/fes/prototype"
-  ]
-	settings:
-	  core:
-	    "file-types":
-        'html': 'text.html.nunjucks'
-}
